career-app/requirements.txt
torch
unsloth
datasets
trl
transformers
fastapi
uvicorn
python-multipart
PyMuPDF
python-docx